show dishes in the cart or remove dish from the cart
1. Show all dishes in a shopping cart. Input is a cart ID. First checks whether that cart ID is valid. If not print a message invalid cart ID. If the ID is valid, print out every dish in 
the shopping cart, including dish name, price, quantity.
2. Remove a dish from shopping cart. Input includes dish ID and cart ID. First check whether the cart with the given ID has that dish. If not print a message 'Invalid 
input'. If the input ID is valid, check the quantity of that dish. If it is more than one, then reduce the quantity of that dish from the cart and print a message saying ‘quantity reduced’. If the quantity is one, delete that row from the cart and print out 'dish removed'.