Update status of an order.
Update status of an order. Input is order ID, new status (1 is in progress, 2 is delivered, 3 is canceled), and input time. The procedure does the following:
1) First checks whether the order ID is valid. If not print a message saying invalid order id. 
2) Update the status of the order to the input status. In case new status is in progress, no additional action is needed. 
3) In case new status is ‘delivered’, insert a message into message table for the corresponding customer, with message time as input time, and message body saying 'Your 
order X has been delivered!' where X is the order ID. 
4) In case new status is ‘canceled’, update the status to canceled, insert a message into message table for the corresponding customer, with message time as input time, and 
message body saying 'Your order X has been canceled and refund issued!' where X is the order ID. Please also insert into payment table a refund record with a new payment ID, the corresponding customer id and order id, time as input time, and amount as the negative of the total amount in the order, and payment method the same as the original payment record.