Compute total amount for dishes in a cart. Input: cart ID and a checkout time, delivery method (1 deliver, 2 pickup). The procedure first checks whether the cart ID is valid. It then computes the total
1) sum up price * quantity of each dish in the cart 
2) discount is then applied for the customer if the discount is still valid at checkout time (check the discount start and end time associated with the customer). 
For fixed percentage discount, the sum in step 1) is multiplied by 1-discount rate. E.g., 10% discount means the price should be multiplied by 0.9. For fixed amount discount, the sum is deducted by that amount. Free delivery discount will be handled in step 3). 
3) delivery fee is added to the result in step 2 if the delivery method is not pickup. If the restaurant is at the same zip code as the customer's home address, the delivery fee is $2.00. If the zip code is different, the delivery fee is $5.00. The delivery fee is zero if a free delivery
discount is valid for the customer at time of order.